mGetModuleAddress function and its overload should now properly return 64 bit addresses

+++ MemTools/mProcess.cpp
@@ -49,7 +49,7 @@ namespace mProcessFunctions {
 		HANDLE targetProcessHandle = mGetHandle(PROCESS_NAME, ProcessAccess::QueryInformation);
 		HMODULE hMods[1024];
 		DWORD cbNeeded;
-		DWORD targetModuleAddress = NULL;
+		uintptr_t targetModuleAddress = NULL;
 
 		if (targetProcessHandle != NULL) {
 			if (EnumProcessModules(targetProcessHandle, hMods, sizeof(hMods), &cbNeeded)) {
@@ -57,7 +57,7 @@ namespace mProcessFunctions {
 					char szModName[MAX_PATH];
 					if (GetModuleBaseNameA(targetProcessHandle, hMods[i], szModName, sizeof(szModName) / sizeof(TCHAR))) {
 						if (strcmp(szModName, MODULE_NAME.c_str()) == 0) {
-							targetModuleAddress = (DWORD)hMods[i];
+							targetModuleAddress = (uintptr_t)hMods[i];
 							break;
 						}
 					}
@@ -69,17 +69,17 @@ namespace mProcessFunctions {
 		return targetModuleAddress;
 	}
 
-	DWORD mGetModuleAddress(const HANDLE &PROCESS_HANDLE, const std::string &MODULE_NAME) {
+	uintptr_t mGetModuleAddress(const HANDLE &PROCESS_HANDLE, const std::string &MODULE_NAME) {
 		HMODULE hMods[1024];
 		DWORD cbNeeded;
-		DWORD targetModuleAddress = NULL;
+		uintptr_t targetModuleAddress = NULL;
 
 		if (EnumProcessModules(PROCESS_HANDLE, hMods, sizeof(hMods), &cbNeeded)) {
 			for (unsigned int i = 0; i < (cbNeeded / sizeof(HMODULE)); i++) {
 				char szModName[MAX_PATH];
 				if (GetModuleBaseNameA(PROCESS_HANDLE, hMods[i], szModName, sizeof(szModName) / sizeof(TCHAR))) {
 					if (strcmp(szModName, MODULE_NAME.c_str()) == 0) {
-						targetModuleAddress = (DWORD)hMods[i];
+						targetModuleAddress = (uintptr_t)hMods[i];
 						break;
 					}
 				}
